* EmbeddedSoftware/AVR/avrlib: Fixed 16 bit truncation of time in timebase.c. Removed some output files. Small changes in Makefile and ldscripts.
git-svn-id: https://svn.arune.se/svn/HomeAutomation@362 e69ec2e4-db17-0410-bd4a-d6f28f89cd0d

+++ trunk/EmbeddedSoftware/AVR/avr-lib/drivers/timer/timebase.c
@@ -106,7 +106,7 @@ void Timebase_Init() {
  */
 uint32_t Timebase_CurrentTime(void) {
 	uint8_t sreg;
-	uint16_t res;
+	uint32_t res;
 	sreg=SREG;
 	cli();
 	res = gMilliSecTick;
@@ -125,10 +125,10 @@ uint32_t Timebase_CurrentTime(void) {
  */ 
 uint32_t Timebase_PassedTimeMillis(uint32_t t0) {
 	uint8_t sreg;
-	uint16_t res;
+	uint32_t res;
 	sreg=SREG;
 	cli();
-	res = (uint16_t)(gMilliSecTick-t0);
+	res = (uint32_t)(gMilliSecTick-t0);
 	SREG=sreg;
 	return res;
 }
